Responisve design for tablets on index.html
Added responsive design for small tablets with width in the range of 568px to 768px and for big tablets with width in the range of 768px to 1024px
Started working on responsive design for laptops. Added asset for main carousel - large image for "use our booking service" slide (register-now-large.png).