Booking design finished; asset color updated
The overall design of the booking calendar is finished (including responsiveness).
It’s designed to handle up to five appointments on a single screen (limits will be enforced via JavaScript).
Updated the color of the "cross.svg" asset, which is used in the booking calendar and will also be used in the mobile pop-up menu.